tsnnls now builds on windows with msvc.  This has been tested with msvc 8 (2005).
A win32 directory has been added, which contains the project file and solution file, as well as the config file.  This config file was generated by hand, and could probably be simplified.  

I wasn't sure how to add to the autoconf system, so when there was a symbol or function missing on windows, I erred on the side of reducing functionality.  I added some ifdefs around rusage, which disable the timing in tsnnls_test on windows. I switched snprintf to sprintf with an assert, and I replaced setenv with putenv.  The changes appear to work on gcc and msvc.  If we want the functionality back, it's probably just a matter of updating autoconf appropriately.

The biggest change I made to get everything compiling on msvc was to change the compilation mode to c++ for several of the c file (only in the msvc project file).  gcc supports C99, but msvc does not.  It ended up being simpler to convert the c to valid c++ than to the older dialect of c, and much less error-prone - just had to cast memory explicitly.

For now, the blas and lapack libraries are not included in the repository, because they are 29MB total, and might be annoying.  I figured I'd ask first.  Instructions for obtaining working libraries are included.